refactor(chrome-extension): split content.js into layered modules
Phase 2 of the architecture cleanup. Carve the 2.6k-line content.js
monolith along its existing seams into three framework-free declaration
modules, leaving content.js as the orchestration/IPC layer (down from
2566 to 976 lines):

- parsing.js      pure parsing/formatting helpers (range expressions,
                  date/time detection + rounding, in-text number
                  extraction, numeric formatting)
- dom-adapters.js the TableAdapter abstraction (NativeTableAdapter /
                  GridAdapter / makeAdapter) plus the structure-preserving
                  cell-write primitives and grid-shape constants
- ui-toggle.js    the per-table toggle widget UI (geometry, styles,
                  button creation/positioning, highlight/flash) and the
                  table/grid detection predicates

All load-time-executing code (context-menu/message/pagehide listeners and
the MutationObserver wiring) stays in content.js, which loads last. Pure
file moves; no behavior change.

Load order kept in lockstep across manifest content_scripts, the test
harness eval concatenation, and the secondary sandbox eval; guard tests
assert the ordering and that content.js loads last. Source-scan tests now
read the combined content-script source so they are location-agnostic.

node chrome-extension/tests.js (955) and node js/tests.js (158) pass.
